pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "modmail-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Discord
    "discord.py>=2.3.2",
    # Async HTTP (LLM providers)
    "httpx>=0.27.0",
    # Data validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Database
    "asyncpg>=0.29.0",
    # Cache + queues
    "redis[asyncio]>=5.0.0",
    "arq>=0.26.0",
    # Logging
    "structlog>=24.2.0",
    # LLM providers (optional, install what you use)
    "anthropic>=0.28.0",
    "openai>=1.35.0",
    # Utils
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
